Move AspirinListener interface into the listener package.
