+++ src/main/java/org/masukomi/aspirin/core/listener/AspirinListener.java
@@ -20,7 +20,7 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY
 WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.masukomi.aspirin.core;
+package org.masukomi.aspirin.core.listener;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
